Working on group parsing for bib2table.py
Working on creating a structure that creates markdown tables for each group category with the tags organised.

Pandas is being used for ease of creation of markdown tables here but the goal is to only use that for the output so that other things can be slot in the future (same idea for pybtex).

Currently the tables have not been created but will be done soon.